backend/requirements.txt
fastapi
uvicorn
pyserial
requests
sqlalchemy
solana
solders
python-dotenv